+++ example/flutter/lib/models/message.dart
@@ -1,9 +1,0 @@
-import 'package:mqtt_client/mqtt_client.dart' as mqtt;
-
-class Message {
-  final String topic;
-  final String message;
-  final mqtt.MqttQos qos;
-
-  Message({this.topic, this.message, this.qos});
-}